moves class to file console => extra
