+++ lib/Extension/ClassToFileExtra/Command/FileInfoCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phpactor\Extension\ClassToFileConsole\Command;
+namespace Phpactor\Extension\ClassToFileExtra\Command;
 
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -8,7 +8,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Input\InputArgument;
 use Phpactor\Extension\Core\Console\Dumper\DumperRegistry;
 use Phpactor\Extension\Core\Console\Handler\FormatHandler;
-use Phpactor\Extension\ClassToFileConsole\Application\FileInfo;
+use Phpactor\Extension\ClassToFileExtra\Application\FileInfo;
 
 class FileInfoCommand extends Command
 {
